+++ CHANGELOG.md
@@ -5,7 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: `TextInputWithUnit`: Add `secondUnit` prop.
-- Feature: `HorizontalCrowdfundingCard`: Add `noProgress` prop, defaults to `false`.
+
+## [7.1.0] - 2022-01-20
+
+Features: 
+
+- `HorizontalCrowdfundingCard`: Add `noProgress` prop, defaults to `false`.
+- `TextCopy`: Add `size` prop, default to `regular`.
 
 ## [7.0.0] - 2022-01-18